VSE/requirements.txt
flask>=3.0.0
requests>=2.31.0
pywebview>=5.0.0
pyinstaller>=6.0.0
